+++ src/parser/spells/activation.js
@@ -4,13 +4,15 @@ import DICTIONARY from "../../dictionary.js";
  * Gets the activation information of this spell
  */
 export function getActivation(data) {
+  // for newer override spells, activation is at higher level
+  const activation = data.activation ? data.activation : data.definition.activation;
   const activationType = DICTIONARY.spell.activationTypes.find(
-    (type) => type.activationType === data.definition.activation.activationType
+    (type) => type.activationType === activation.activationType
   );
-  if (activationType && data.definition.activation.activationTime) {
+  if (activationType && activation.activationTime) {
     return {
       type: activationType.value,
-      cost: data.definition.activation.activationTime,
+      cost: activation.activationTime,
       condition: data.definition.castingTimeDescription || "",
     };
   } else {