+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.9.58
+
+* Spells with limited uses based on proficiency now render uses correctly.
+* Some spells from races would not parse the duration correctly if different from original spell.
+
 # 2.9.57
 
 * Feature effect: Rage created a damage field when it shouldn't.